Kaggle/ragChatbot/requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-community==0.0.13

# Vector Store
pinecone-client==3.0.0

# Embeddings
openai==1.7.0

# LLM
anthropic==0.8.0

# Document Loaders
pypdf==3.17.4

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Optional but recommended
pandas==2.1.4
numpy==1.26.2
